Move some args to common
